feat: func run support all build options
Refactors the run command to include all options from build, in the same
manner as deploy.
Updates help text to match the other command format, and lays the verbal
groundwork for the difference between containerized (currently default)
and non-containerized runs.
Adds the --container flag (currently defaults to true) which will, when
disabled, invoke the localhost-based runner which, in this commit, is
not yet active.